Restructure `services/ethereum`
