+++ src/renderer/contexts/EthereumContext.tsx
@@ -1,9 +1,6 @@
 import React, { createContext, useContext } from 'react'
 
-import { useSubscription } from 'observable-hooks'
-
-import { client$, setNetworkState, address$ } from '../services/ethereum/service'
-import { useAppContext } from './AppContext'
+import { client$, address$ } from '../services/ethereum'
 
 export type EthereumContextValue = {
   client$: typeof client$
@@ -18,8 +15,6 @@ const initialContext: EthereumContextValue = {
 const EthereumContext = createContext<EthereumContextValue | null>(null)
 
 export const EthereumProvider: React.FC = ({ children }): JSX.Element => {
-  const { network$ } = useAppContext()
-  useSubscription(network$, (network) => setNetworkState(network))
   return <EthereumContext.Provider value={initialContext}>{children}</EthereumContext.Provider>
 }
