+++ src/renderer/services/ethereum/index.ts
@@ -1,0 +1,4 @@
+import { balances$, reloadBalances } from './balances'
+import { client$, address$ } from './common'
+
+export { client$, address$, reloadBalances, balances$ }
